Update build.gradle dependency to use Boot version
Update the build.gradle dependency for
io.spring.dependency-management to not specifiy the version.
This will cause the version that Spring Boot depends on to
be used. This is preferred because when dependabot is used in
the project we can configure it to exclude all patch release versions.
The dependency-management version does not follow the Boot versions,
so having the dependency managed automatically is cleaner for
the dependabot config.